Restructure Client-side
