+++ client/src/context/contextProvider.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const contextProvider = () => {
+    return (
+        <div>
+            
+        </div>
+    );
+};
+
+export default contextProvider;
