+++ client/src/pages/Login/connexion_main.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import "../styles/Connexion_main.css";
+import "./connexion_main.css";
 import { Link } from "react-router-dom";
 import Button from '@mui/material/Button';
